refactor: split base props file
