+++ packages/x-data-grid/src/models/gridSlotsComponentsProps.ts
@@ -33,33 +33,10 @@ import type { GridColumnsManagementProps } from '../components/columnsManagement
 import type { GridLoadingOverlayProps } from '../components/GridLoadingOverlay';
 import type { GridRowCountProps } from '../components/GridRowCount';
 import type { GridColumnHeaderSortIconProps } from '../components/columnHeaders/GridColumnHeaderSortIcon';
+import type { BadgeProps, DividerProps, MenuItemProps } from './gridBaseSlots';
 
 type RootProps = React.HTMLAttributes<HTMLDivElement> & Record<`data-${string}`, string>;
 type MainProps = React.HTMLAttributes<HTMLDivElement> & Record<`data-${string}`, string>;
-
-export type BadgeProps = {
-  badgeContent?: React.ReactNode;
-  children: React.ReactNode;
-  color?: 'primary' | 'default' | 'error';
-  overlap?: 'circular';
-  variant?: 'dot';
-  invisible?: boolean;
-};
-
-export type DividerProps = {};
-
-export type MenuItemProps = {
-  autoFocus?: boolean;
-  children: React.ReactNode;
-  /** For items that aren't interactive themselves (but may contain an interactive widget) */
-  inert?: boolean;
-  disabled?: boolean;
-  onClick?: React.MouseEventHandler<HTMLElement>;
-  iconStart?: React.ReactNode;
-  iconEnd?: React.ReactNode;
-  selected?: boolean;
-  value?: number | string | readonly string[];
-};
 
 // Overrides for module augmentation
 export interface BaseBadgePropsOverrides {}
