+++ scripts/x-data-grid-pro.exports.json
@@ -518,7 +518,7 @@
   { "name": "GridSkeletonCell", "kind": "Variable" },
   { "name": "GridSkeletonCellProps", "kind": "Interface" },
   { "name": "GridSkeletonRowNode", "kind": "Interface" },
-  { "name": "GridSlotProps", "kind": "Interface" },
+  { "name": "GridSlotProps", "kind": "TypeAlias" },
   { "name": "GridSlots", "kind": "Interface" },
   { "name": "GridSlotsComponent", "kind": "Interface" },
   { "name": "GridSlotsComponentsProps", "kind": "TypeAlias" },
