+++ scripts/x-data-grid.exports.json
@@ -469,7 +469,7 @@
   { "name": "GridSkeletonCell", "kind": "Variable" },
   { "name": "GridSkeletonCellProps", "kind": "Interface" },
   { "name": "GridSkeletonRowNode", "kind": "Interface" },
-  { "name": "GridSlotProps", "kind": "Interface" },
+  { "name": "GridSlotProps", "kind": "TypeAlias" },
   { "name": "GridSlots", "kind": "Interface" },
   { "name": "GridSlotsComponent", "kind": "Interface" },
   { "name": "GridSlotsComponentsProps", "kind": "TypeAlias" },
